--- src/options/index.js
@@ -1,2 +1,0 @@
-export * from "./buildOptions";
-export * from "./storybookOptions";
